fix #9: feat(auth): user auth close #9
