--- src.bak/app/Controller/Admin/AppController.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace App\Controller\Admin;
-
-class AppController extends \App\Controller\AppController
-{
-    protected $template;
-    
-    public function __construct()
-    {
-        parent::__construct();
-        if(!$this->auth->isAdmin())
-        {
-            header('Location: ?p=error.forbidden');
-        }
-    }
-
-}
