--- src.bak/app/Controller/PageController.php
@@ -1,54 +1,0 @@
-<?php
-namespace App\Controller;
-use Core\Html\Form;
-use Core\Mail\Mail;
-
-class PageController extends AppController
-{
-    public function __construct()
-    {
-        parent::__construct();
-    }
-
-    public function contact()
-    {
-        if(!empty($_POST))
-        {
-            if (empty($_POST['name']))
-            {
-                $_SESSION['flash']['error'] = 'Veuillez renseigner votre nom';
-            }    
-            elseif (empty($_POST['email']) || !filter_var($_POST['email'], FILTER_VALIDATE_EMAIL))
-            {
-                $_SESSION['flash']['error'] = 'Veuillez renseigner un email valide';
-            } 
-            elseif (empty($_POST['subject']))
-            {
-                $_SESSION['flash']['error'] = 'Veuillez renseigner le sujet de votre message';
-            } 
-            elseif (empty($_POST['message']))
-            {
-                $_SESSION['flash']['error'] = 'Veuillez renseigner votre message';
-            }
-            else
-            {
-                $to = 'contact@axelbuob.fr';
-                $subject = $_POST['subject'];
-                $message = $_POST['message'];
-                $from = $_POST['from'];
-                $mail = new Mail($to, $from, $subject, $message);
-                $mail->send();
-                $_SESSION['flash']['success'] = 'Merci votre email a bien été envoyé';
-                header('Location: ?p=post.index');
-                die();
-            }
-        }
-        $form = new Form();
-        $this->render('page.contact', compact('form'));
-    }
-
-    public function resume()
-    {
-        $this->render('page.resume');
-    }
-}
