--- src.bak/app/Entity/CategoryEntity.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace App\Entity;
-
-class CategoryEntity extends \Core\Entity\Entity
-{
-    public function getUrl()
-    {
-        return '?p=post.category&id=' . $this->id;
-    }
-
-}
