--- src.bak/app/Entity/UserEntity.php
@@ -1,6 +1,0 @@
-<?php
-namespace App\Entity;
-
-class UserEntity extends \Core\Entity\Entity
-{
-}
