--- src.bak/app/Table/CategoryTable.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace App\Table;
-
-class CategoryTable extends \Core\Table\Table
-{
-    
-}
