--- src.bak/app/view/error/403.php
@@ -1,1 +1,0 @@
-<h1>403 Forbidden</h1>
