--- src.bak/app/view/error/404.php
@@ -1,1 +1,0 @@
-<h1>404 Not Found</h1>
