--- src.bak/app/view/page/resume.php
@@ -1,1 +1,0 @@
-<h1>Resume</h1>
