--- src.bak/app/view/user/confirm.php
@@ -1,2 +1,0 @@
-<p><?= $message; ?></p>
-<a href="index.php">Retour à l'accueil</a>
