--- src.bak/app/view/user/signin.php
@@ -1,7 +1,0 @@
-<h1>Connection</h1>
-<form method="post">
-    <?= $form->input("email", "Email", "Votre email"); ?>
-    <?= $form->input("password", "Mot de passe", "Votre mot de passe", ["type" => "password"]); ?>
-    <?= $form->submit("Envoyer"); ?>
-    <a href="?p=user.forget">Réinitialiser le mot de passe</a>
-</form>
