--- src.bak/core/Html/Text.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Core\Html;
-
-class Text
-{
-    
-}
